backend/ai-service/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
google-generativeai==0.3.1
supabase==2.0.3
numpy==1.26.2
scipy==1.11.4
pydantic==2.5.2
